update ndef and pid to better support x-plat + litgen
